WIP improving test tooling
I'm trying to write the tests for the changes to EIP editor handling. But the repository lacks necessary test tooling that would make things easier. Maybe there's a package that could make this easier? Either way, I'm stuck on trying to make jest mocks work with types. I'm pretty sure it's possible, but I think I'll push this problem off to later because I'm wasting a lot of time on it.

It's worth having good tooling though because it will streamline tests in the future; making them faster to make and easier to maintain. Which tends to be worth the investment.