cleanup the code and better message on new eip
previously, new files for eips (i.e. if they didn't have eip number assigned yet) got an error about how the eip fielname was not formatted correct.

this was because it was instructed that the user provide a {summary of eip}.md filename.

To avoid confusing new authors, this commit instead explains that the eip needs an eip number and it mentions the editors