added maintainers as parameter
This commit adds maintainers to the configuration. So that one can change who is mentioned when issues arise.

While I was building this I got a little into the weeds wrt error boundary handling. I did this mostly because it's important for usability that errors be handled consistently, but I also spent time because it's a catch 22 that you would want to mention the maintainers on an error and yet the input could have an error itself.

To avoid this conflict I simply built in error handling for getting the maintainers that would then propose an error string if something was off. This seems to work well and was a minimal amount of extra work to implement.