Trying to fix auto-detection.
The tool that GitHub uses to match licenses allegedly strips all whitespace, so this should be an exact match.  If it isn't we should figure out what is different between the two.  I added a line at a top here which is obviously different between the two, lets see if that helps.